requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
Django==1.11.16
djangorestframework==3.7.7
Faker==0.9.1
idna==2.7
mixer==6.1.3
more-itertools==4.3.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
pytest==3.8.2
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
